fix(auto-expand): retry sibling detection at post-processing time (#45)
User's HorribleSubs JoJo Stardust 1-48 grab lost Egypt-hen because
grab-time auto-expand bailed when qBit's metadata wait timed out after
60s. Without routes written, post-processing fell through to the
legacy single-series path and pushed all 48 files into Stardust's
folder; Egypt-hen was never auto-added.

Two changes to make grab-time metadata failures non-fatal:

1. Extract the core expansion into `services::auto_expand` so it's
   callable from both handler and service layers. Keeps the outer
   grab-time wrapper in `handlers::library::search` for the qBit
   metadata wait; the inner `expand_from_files` is pure and shareable.

2. Retry in `services::post_processing::import_torrent`: when `routes`
   is empty and the grab is a batch, call `expand_from_files` with
   the file list qBit just returned (available by definition at
   import time — we're about to move the files). If it detects
   siblings now, route rows get written and the regular import loop
   picks them up.

Also bumped the grab-time metadata-wait timeout from 60s to 180s;
slow-DHT magnets or public trackers under load need more than a
minute for qBit to fetch metadata. The retry fallback catches the
long-tail cases where even 180s isn't enough.

Test: `expand_from_files_rescues_grab_when_called_at_post_processing_time`
simulates the exact user scenario (parent in DB, grab row exists, no
routes, 48-ep JoJo pack) and verifies the retry adds Egypt-hen with
correct offset/route after grab-time expansion failed.
